option to NOT process flip of logical state
to use the same push button object in menu and as alternative logical switch (when menu is disabled) I need the possibility to ignore / not do the flip of the logical button state after a debounced press is detected.

default is flipping. Optionally call processButtonState(false) to skip the flipping.